+++ coaster/utils.py
@@ -43,17 +43,18 @@ _tag_re = re.compile('<.*?>')
 def buid():
     """
     Return a new random id that is exactly 22 characters long,
-    by encoding a UUID4 in URL-safe Base64. See
+    by encoding a UUID1MC in URL-safe Base64. See
     http://en.wikipedia.org/wiki/Base64#Variants_summary_table
 
-    >>> len(newid())
+    >>> len(buid())
     22
-    >>> newid() == newid()
+    >>> buid() == buid()
     False
-    >>> isinstance(newid(), unicode)
+    >>> isinstance(buid(), unicode)
     True
     """
-    return unicode(urlsafe_b64encode(uuid.uuid4().bytes).rstrip('='))
+    return unicode(urlsafe_b64encode(uuid1mc().bytes).rstrip('='))
+
 
 # Retain old name
 newid = buid