+++ CHANGES.rst
@@ -5,7 +5,7 @@
 * Deprecated ``add_and_commit`` in favour of ``failsafe_add``.
 * New utils: ``uuid2buid`` and ``buid2uuid``.
 * Removed ``timestamp_columns`` (was deprecated in 0.4.3).
-* Replaced ``py-bcrypt`` dependency with ``python-bcrypt``.
+* Replaced ``py-bcrypt`` dependency with ``bcrypt``.
 
 0.5.0
 -----